+++ examples/express-api-reference/src/index.ts
@@ -49,6 +49,6 @@ app.use(
 )
 
 // Listen
-app.listen(5056, () => {
-  console.log('💻 Express listening on http://localhost:5056')
+app.listen(5055, () => {
+  console.log('💻 Express listening on http://localhost:5055')
 })