+++ packages/api-client/src/components/ApiClient/Response/ResponseBody.vue
@@ -1,5 +1,5 @@
 <script lang="ts" setup>
-import { CodeMirror, type CodeMirrorLanguage } from '@scalar/use-codemirror'
+import { ScalarCodeBlock } from '@scalar/components'
 import contentType from 'content-type'
 import { computed } from 'vue'
 
@@ -33,8 +33,12 @@ const mediaType = computed(() => {
   }
 })
 
-const codeMirrorLanguage = computed((): CodeMirrorLanguage | null => {
-  if (mediaType.value === 'application/json') {
+const codeMirrorLanguage = computed((): string | null => {
+  if (
+    mediaType.value === 'application/json' ||
+    mediaType.value === 'application/problem+json' ||
+    mediaType.value === 'application/vnd.api+json'
+  ) {
     return 'json'
   }
 
@@ -52,11 +56,11 @@ const codeMirrorLanguage = computed((): CodeMirrorLanguage | null => {
 <template>
   <CollapsibleSection title="Body">
     <template v-if="active">
-      <CodeMirror
+      <ScalarCodeBlock
         v-if="codeMirrorLanguage"
+        class="custom-scroll"
         :content="data"
-        :language="codeMirrorLanguage"
-        readOnly />
+        :lang="codeMirrorLanguage" />
       <div
         v-else
         class="scalar-api-client__empty-state">