+++ packages/api-reference/src/components/SearchButton.vue
@@ -32,7 +32,8 @@ whenever(keys[`meta_${props.searchHotKey}`], () =>
     @click="modalState.show">
     <ScalarIcon
       class="search-icon"
-      icon="Search" />
+      icon="Search"
+      size="sm" />
     <div class="sidebar-search-input">
       <span class="sidebar-search-placeholder">Search</span>
       <span class="sidebar-search-shortcut">
@@ -51,8 +52,8 @@ whenever(keys[`meta_${props.searchHotKey}`], () =>
   display: flex;
   align-items: center;
   position: relative;
-  padding: 0 3px 0 12px;
-  min-width: 224px;
+  padding: 0 3px 0 9px;
+  min-width: 254px;
   max-width: 100%;
   font-family: var(--theme-font, var(--default-theme-font));
   background: var(
@@ -118,7 +119,7 @@ whenever(keys[`meta_${props.searchHotKey}`], () =>
 
 .search-icon {
   padding: 0;
-  margin-right: 9px;
+  margin-right: 6px;
   width: 12px;
 }
 </style>