+++ packages/api-reference/vite.config.ts
@@ -1,6 +1,5 @@
 import vue from '@vitejs/plugin-vue'
 import path from 'path'
-import { webpackStats } from 'rollup-plugin-webpack-stats'
 import cssInjectedByJsPlugin from 'vite-plugin-css-injected-by-js'
 import { defineConfig } from 'vitest/config'
 
@@ -10,7 +9,7 @@ export default defineConfig({
   define: {
     'process.env.NODE_ENV': '"production"',
   },
-  plugins: [vue(), cssInjectedByJsPlugin(), webpackStats()],
+  plugins: [vue(), cssInjectedByJsPlugin()],
   build: {
     emptyOutDir: true,
     cssCodeSplit: false,
@@ -37,7 +36,7 @@ export default defineConfig({
         // Resolve the uncompiled source code for all @scalar packages
         // @scalar/* -> packages/*/
         // (not @scalar/components/*/style.css)
-        find: /^@scalar\/(?!(snippetz|components\/style\.css|components|themes\b))(.+)/,
+        find: /^@scalar\/(?!(openapi-parser|snippetz|components\/style\.css|components\b))(.+)/,
         replacement: path.resolve(__dirname, '../$2/src/index.ts'),
       },
     ],