+++ packages/components/vite.config.ts
@@ -2,6 +2,7 @@ import vue from '@vitejs/plugin-vue'
 import { URL, fileURLToPath } from 'node:url'
 import * as path from 'path'
 import { defineConfig } from 'vite'
+import cssInjectedByJsPlugin from 'vite-plugin-css-injected-by-js'
 import dts from 'vite-plugin-dts'
 
 import pkg from './package.json'
@@ -13,7 +14,7 @@ export default defineConfig({
       // Could also be a dictionary or array of multiple entry points
       entry: './src/index.ts',
       name: '@scalar/components',
-      formats: ['es', 'cjs', 'umd'],
+      formats: ['es', 'cjs'],
       fileName: 'index',
       // the proper extensions will be added
       // fileName: 'my-lib',
@@ -30,20 +31,11 @@ export default defineConfig({
           (item) => !item.startsWith('@scalar'),
         ),
       ],
-      output: {
-        // Provide global variables to use in the UMD build
-        // for externalized deps
-        exports: 'named',
-        globals: {
-          vue: 'Vue',
-        },
-      },
     },
   },
-  plugins: [vue(), dts({ insertTypesEntry: true, rollupTypes: true })],
-  resolve: {
-    alias: {
-      '@': fileURLToPath(new URL('./src', import.meta.url)),
-    },
-  },
+  plugins: [
+    vue(),
+    dts({ insertTypesEntry: true, rollupTypes: true }),
+    cssInjectedByJsPlugin(),
+  ],
 })